# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API สำหรับระบบจัดการข้อมูลผู้ป่วย",
        "title": "Agnos Patient API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/patient/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "สร้างผู้ป่วยใหม่ภายใต้โรงพยาบาลของ staff ที่ login อยู่",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "สร้างข้อมูลผู้ป่วยใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลผู้ป่วย",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PatientInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายใน",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/patient/search": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ค้นหาผู้ป่วยภายใต้โรงพยาบาลของ staff ผ่าน filter ต่างๆ (ทุก field เป็น optional)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "ค้นหาผู้ป่วย",
                "parameters": [
                    {
                        "description": "เงื่อนไขการค้นหา (ทุก field optional)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PatientSearchInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "รายการผู้ป่วยที่พบ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายใน",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/patient/search/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ค้นหาผู้ป่วยด้วย National ID หรือ Passport ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "ค้นหาผู้ป่วยด้วย ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "National ID หรือ Passport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ข้อมูลผู้ป่วย",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ไม่ได้ระบุ ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "ไม่พบข้อมูล",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/staff/create": {
            "post": {
                "description": "ลงทะเบียน staff พร้อมระบุโรงพยาบาล (ถ้าโรงพยาบาลยังไม่มีจะสร้างให้อัตโนมัติ)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "สร้าง staff ใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูล Staff",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateStaffInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "สร้าง staff สำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายใน",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/staff/login": {
            "post": {
                "description": "ตรวจสอบ username/password แล้วคืน JWT access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "เข้าสู่ระบบ",
                "parameters": [
                    {
                        "description": "ข้อมูล Login",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateStaffInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "access_token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "username หรือ password ไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreateStaffInput": {
            "type": "object",
            "required": [
                "hospital_id",
                "password",
                "username"
            ],
            "properties": {
                "hospital_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "domain.Gender": {
            "type": "string",
            "enum": [
                "m",
                "f",
                "o"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderOther"
            ]
        },
        "domain.PatientInput": {
            "type": "object",
            "required": [
                "first_name_en",
                "first_name_th",
                "last_name_en",
                "last_name_th",
                "national_id",
                "patient_hn"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name_en": {
                    "type": "string"
                },
                "first_name_th": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/domain.Gender"
                },
                "last_name_en": {
                    "type": "string"
                },
                "last_name_th": {
                    "type": "string"
                },
                "middle_name_en": {
                    "type": "string"
                },
                "middle_name_th": {
                    "type": "string"
                },
                "national_id": {
                    "type": "string"
                },
                "passport_id": {
                    "type": "string"
                },
                "patient_hn": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "domain.PatientSearchInput": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name_en": {
                    "type": "string"
                },
                "first_name_th": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/domain.Gender"
                },
                "last_name_en": {
                    "type": "string"
                },
                "last_name_th": {
                    "type": "string"
                },
                "middle_name_en": {
                    "type": "string"
                },
                "middle_name_th": {
                    "type": "string"
                },
                "national_id": {
                    "type": "string"
                },
                "passport_id": {
                    "type": "string"
                },
                "patient_hn": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ใส่ token ในรูปแบบ: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}